% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje.R
\name{analizuj_zarobki}
\alias{analizuj_zarobki}
\title{Analiza zarobków}
\usage{
analizuj_zarobki(dane, wybrane_stanowisko)
}
\arguments{
\item{dane}{data frame zawierający kolumny stanowisko i zarobki}

\item{wybrane_stanowisko}{nazwa stanowiska (np. "Developer")}
}
\value{
obiekt klasy \code{analiza_zarobkow} (data.frame ze statystykami:
średnia, minimum i maksimum)
}
\description{
Funkcja filtruje dane według stanowiska i oblicza podstawowe statystyki wynagrodzeń.
}
\examples{
analizuj_zarobki(dane_it, "Developer")
}
