% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_mse.R
\name{calculate_mse}
\alias{calculate_mse}
\title{Obliczanie średniego błędu kwadratowego (MSE)}
\usage{
calculate_mse(actual, predicted)
}
\arguments{
\item{actual}{Wektor rzeczywistych wartości.}

\item{predicted}{Wektor przewidywanych wartości.}
}
\value{
Średni błąd kwadratowy (MSE).
}
\description{
Ta funkcja oblicza średni błąd kwadratowy między rzeczywistymi a przewidywanymi wartościami.
}
