% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_mse.R
\name{normalize_mse}
\alias{normalize_mse}
\title{Normalizacja średniego błędu kwadratowego (MSE)}
\usage{
normalize_mse(actual, predicted, other_predicted)
}
\arguments{
\item{actual}{Wektor rzeczywistych wartości.}

\item{predicted}{Wektor przewidywanych wartości.}

\item{other_predicted}{Wektor przewidywanych wartości dla porównania.}
}
\value{
Znormalizowany średni błąd kwadratowy (MSE).
}
\description{
Ta funkcja normalizuje średni błąd kwadratowy (MSE) między dwoma zestawami przewidywań.
}
