% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictions.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Wykresowanie rzeczywistych i przewidywanych cen akcji}
\usage{
plot_predictions(model, stock_prices)
}
\arguments{
\item{model}{Model predykcji do użycia. Dostępne opcje to: "knn", "arima", "prophet", "lm", "ma".}

\item{stock_prices}{Wektor liczbowy zawierający ceny akcji. Wektor powinien zawierać co najmniej 100 elementów, z czego pierwsze 90 elementów jest używane do treningu modelu, a ostatnie 10 do testowania.}
}
\value{
Funkcja nie zwraca żadnej wartości. Generuje wykres rzeczywistych i przewidywanych cen akcji oraz wyświetla błędy.
}
\description{
Ta funkcja tworzy wykres rzeczywistych i przewidywanych cen akcji, wykorzystując różne modele predykcji. Na wykresie przedstawione są rzeczywiste ceny akcji, przewidywane ceny akcji przez wybrany model oraz regresję liniową. Dodatkowo obliczane i wyświetlane są błędy dla wybranego modelu i regresji liniowej.
}
\examples{
stock_prices <- c(1.78, 1.7, 1.69, 1.34, 1.05, 1.03, 1, 0.93, 0.94, 1,
                  0.92, 0.92, 0.91, 0.88, 0.9, 0.86, 0.9, 0.98, 0.85, 0.84,
                  0.74, 0.74, 0.78, 0.78, 0.78, 0.79, 0.81, 0.79, 0.79, 0.84,
                  0.83, 0.82, 0.77, 0.75, 0.74, 0.81, 0.82, 0.77, 0.68, 0.7,
                  0.7, 0.75, 0.77, 0.67, 0.74, 0.74, 0.74, 0.76, 0.75, 0.75,
                  0.75, 0.7, 0.69, 0.74, 0.75, 0.75, 0.75, 0.75, 0.75, 0.75,
                  0.75, 0.75, 0.75, 0.76, 0.76, 0.76, 0.76, 0.71, 0.71, 0.71,
                  0.77, 0.77, 0.73, 0.76, 0.76, 0.76, 0.79, 0.8, 0.72, 0.75,
                  0.75, 0.7, 0.75, 0.75, 0.72, 0.72, 0.75, 0.72, 0.72, 0.72,
                  0.67, 0.66, 0.66, 0.63, 0.68, 0.68, 0.61, 0.58, 0.64, 0.65,
                  0.56, 0.46, 0.4, 0.57, 0.5, 0.45, 0.46, 0.46, 0.46, 0.46,
                  0.46, 0.58, 0.58, 0.48, 0.48)
plot_predictions("knn", stock_prices)
plot_predictions("arima", stock_prices)
plot_predictions("prophet", stock_prices)
plot_predictions("lm", stock_prices)
plot_predictions("ma", stock_prices)
}
