% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stock_predictions.R
\name{plot_stock_predictions}
\alias{plot_stock_predictions}
\title{Rysowanie wykresu rzeczywistych i przewidywanych cen akcji}
\usage{
plot_stock_predictions(
  stock_prices,
  predicted_prices,
  predicted_prices_regression,
  mse_model_normalized,
  mse_regression_normalized
)
}
\arguments{
\item{stock_prices}{Wektor rzeczywistych cen akcji.}

\item{predicted_prices}{Wektor przewidywanych cen akcji przez wybrany model.}

\item{predicted_prices_regression}{Wektor przewidywanych cen akcji przez regresję liniową.}

\item{mse_model_normalized}{Znormalizowany średni błąd kwadratowy (MSE) dla wybranego modelu.}

\item{mse_regression_normalized}{Znormalizowany średni błąd kwadratowy (MSE) dla regresji liniowej.}
}
\description{
Ta funkcja tworzy wykres rzeczywistych i przewidywanych cen akcji oraz wyświetla błędy dla wybranego modelu i regresji liniowej.
}
