% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funkcje.R
\name{porownaj_doswiadczenie}
\alias{porownaj_doswiadczenie}
\title{Porównanie doświadczenia}
\usage{
porownaj_doswiadczenie(dane, stanowisko = NULL)
}
\arguments{
\item{dane}{data frame z kolumnami doswiadczenie i zarobki}

\item{stanowisko}{opcjonalny filtr po stanowisku}
}
\value{
obiekt klasy \code{porownanie_doswiadczenia} (data.frame ze
średnimi zarobkami wg doświadczenia)
}
\description{
Funkcja oblicza średnie zarobki dla poziomów doświadczenia.
}
\examples{
porownaj_doswiadczenie(dane_it)
}
