% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_arima.R
\name{predict_arima}
\alias{predict_arima}
\title{Prognozowanie cen akcji za pomocą ARIMA}
\usage{
predict_arima(stock_prices, prange = 90)
}
\arguments{
\item{stock_prices}{Wektor liczbowy zawierający ceny akcji, z których pierwsze 90 elementów jest używane do treningu modelu, a ostatnie 10 do testowania.}
}
\value{
Wektor liczbowy zawierający przewidywane ceny akcji dla ostatnich 10 dni.
}
\description{
Ta funkcja wykorzystuje model ARIMA do przewidywania przyszłych cen akcji na podstawie danych historycznych.
}
\examples{
stock_prices <- c(1.78, 1.7, 1.69, 1.34, 1.05, 1.03, 1, 0.93, 0.94, 1)
predict_arima(stock_prices)
}
