% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_knn.R
\name{predict_knn}
\alias{predict_knn}
\title{Prognozowanie cen akcji za pomocą K-Nearest Neighbors (KNN)}
\usage{
predict_knn(stock_prices)
}
\arguments{
\item{(stock_prices, }{prange) Wektor liczbowy zawierający ceny akcji, z których pierwsze 90 elementów jest używane do treningu modelu, a ostatnie 10 do testowania.
oraz zasięg}
}
\value{
Wektor liczbowy zawierający przewidywane ceny akcji dla ostatnich 10 dni.
}
\description{
Ta funkcja wykorzystuje algorytm K-Nearest Neighbors (KNN) do przewidywania przyszłych cen akcji na podstawie danych historycznych.
}
\examples{
stock_prices <- c(1.78, 1.7, 1.69, 1.34, 1.05, 1.03, 1, 0.93, 0.94, 1)
predict_knn(stock_prices)
}
