% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_sarima.R
\name{predict_sarima}
\alias{predict_sarima}
\title{Prognozowanie cen akcji za pomocą SARIMA}
\usage{
predict_sarima(stock_prices, freq_, p_range = 90)
}
\arguments{
\item{(stock_prices, }{prange) Wektor liczbowy zawierający ceny akcji, z których pierwsze n elementów jest używane do treningu modelu, a ostatnie 10 do testowania
oraz wybrany zasięg}
}
\value{
Wektor liczbowy zawierający przewidywane ceny akcji dla ostatnich 10 dni.
}
\description{
Ta funkcja wykorzystuje model SARIMA do przewidywania przyszłych cen akcji na podstawie danych historycznych.
}
\examples{
stock_prices <- c(1.78, 1.7, 1.69, 1.34, 1.05, 1.03, 1, 0.93, 0.94, 1)
predict_sarima(stock_prices)
}
