% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StockInit.R
\name{StockInit}
\alias{StockInit}
\title{Funkcja wczytuje do programu dane giełdowe z podanego pliku}
\usage{
StockInit(path, cols, range_ = 90)
}
\arguments{
\item{path(string), }{cols(lista stringów), range( x ostatnich dni)}
}
\description{
Funkcja wczytuje do programu dane giełdowe z podanego pliku
}
